boton like arreglado

+++ iTresnaBack/php/senales/post.php
@@ -4,21 +4,19 @@
     $data = json_decode($json);
 
     $nueva_senal = $data->nueva_senal;
-    $desc_senal = $data->desc_senal;
-    $enlace = $data->$enlace;
-    $cod_etiqueta = $date->$cod_etiqueta
-
     $cod_usuario=$data->cod_usuario;
     $cod_org=$data->cod_org;
     $cod_esp=$data->cod_esp;
     $cod_cop=$data->cod_cop;
-    $cod_senal=$data->cod_senal;
 
     $result=array(
         "error"=>1,
         "aniadido"=>0
     );
     if($nueva_senal){
+        $desc_senal = $data->desc_senal;
+        $enlace = $data->$enlace;
+        $cod_etiqueta = $date->$cod_etiqueta;
         if($cod_usuario!=="" && $cod_org!=="" && $cod_esp!=="" && $cod_cop!=="" && $desc_senal!=="" && $enlace!=="" && $cod_etiqueta!==""){
             $sql = "INSERT INTO t_senales 
             VALUES cod_usuario = $cod_usuario 
@@ -42,45 +40,46 @@
 
 
     } else {
+        
+        $cod_senal=$data->cod_senal;
         if($cod_usuario!=="" && $cod_org!=="" 
-        && $cod_esp!=="" && $cod_cop!=="" && $cod_senal!==""){
-            
+            && $cod_esp!=="" && $cod_cop!=="" && $cod_senal!==""){
 
-        if(!haDadoLike($cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal)){
-            $sql="INSERT INTO t_megusta VALUES(?,?,?,?,?)";
-            $query=$conexion->prepare($sql);
-            $query->bind_param("sdddd",$cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal);
-            $query->execute();
-            $affected_rows=$query->affected_rows;
-            $result["error"]=($affected_rows==1)?0:1;
-            $result["aniadido"]=1;
-            $query->close();
-        }else{
-            $sql="DELETE FROM t_megusta WHERE cod_usuario=? AND cod_org=? AND cod_esp=?
-                 AND cod_cop=? AND cod_senal=?";
-            $query=$conexion->prepare($sql);
-            $query->bind_param("sdddd",$cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal);
-            $query->execute();
-            $affected_rows=$query->affected_rows;
-            $result["error"]=($affected_rows==1)?0:1;
-            $query->close();
+            if(!haDadoLike($cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal)){
+                $sql="INSERT INTO t_megusta VALUES(?,?,?,?,?)";
+                $query=$conexion->prepare($sql);
+                $query->bind_param("iiiis",$cod_senal,$cod_cop,$cod_esp,$cod_org,$cod_usuario);
+                $query->execute();
+                $affected_rows=$query->affected_rows;
+                $result["error"]=($affected_rows==1)?0:1;
+                $result["aniadido"]=1;
+                $query->close();
+            }else{
+                $sql="DELETE FROM t_megusta WHERE cod_usuario=? AND cod_org=? AND cod_esp=?
+                    AND cod_cop=? AND cod_senal=?";
+                $query=$conexion->prepare($sql);
+                $query->bind_param("siiii",$cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal);
+                $query->execute();
+                $affected_rows=$query->affected_rows;
+                $result["error"]=($affected_rows==1)?0:1;
+                $result["aniadido"]=0;
+                $query->close();
+            }
         }
     }
+    echo json_encode($result);
     function haDadoLike($cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal):bool{
+        
         include("./../conexion.php");
         $sql="SELECT COUNT(*)
                 FROM t_megusta
                 WHERE cod_usuario=? AND cod_org=? AND cod_esp=?
                 AND cod_cop=? AND cod_senal=?";
         $query=$conexion->prepare($sql);
-        $query->bind_param("sdddd",$cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal);
+        $query->bind_param("siiii",$cod_usuario,$cod_org,$cod_esp,$cod_cop,$cod_senal);
         $query->execute();
         $query->bind_result($cuantos);
         $query->fetch();
         return $cuantos>0;
-
-        
     }
-    echo json_encode($result);
-}
 ?>
